Add and remove iframe event listeners for `create new entry` in preview service
